change type of size arg to randomize to size_t to avoid compiler warnings about signed/unsigned comparison

+++ tests/Naming_Test.cpp
@@ -28,7 +28,7 @@ static char value[BUFSIZ];
 static char type[BUFSIZ];
 
 static void
-randomize (int array[], int size)
+randomize (int array[], size_t size)
 {
   size_t i;
